refactor: separate joi schema into a separate file
