+++ routes/api/contacts.js
@@ -1,32 +1,9 @@
 const express = require("express");
 const { HttpError } = require("../../helpers");
 const contacts = require("../../models/contacts")
-const Joi = require("joi");
+const { userSchema } = require("../../schemas");
 
 const router = express.Router();
-
-// Joi shema for validate req user data
-const userSchema = Joi.object({
-  name: Joi.string().min(3).max(50).required().messages({
-    "string.min": '"name" must be at least 3 characters long',
-    "string.max": '"name" must be at most 50 characters long',
-    "string.empty": '"name" cannot be an empty field',
-    "any.required": '"name" is a required field',
-  }),
-  email: Joi.string().email().required().messages({
-    "string.email": '"email" must be a valid email',
-    "string.empty": '"email" cannot be an empty field',
-    "any.required": '"email" is a required field',
-  }),
-  phone: Joi.string()
-    .pattern(/^\(\d{3}\) \d{3}-\d{4}$/)
-    .required()
-    .messages({
-      "string.pattern.base": '"phone" must be in the format (XXX) XXX-XXXX',
-      "string.empty": '"phone" cannot be an empty field',
-      "any.required": '"phone" is a required field',
-    }),
-});
 
 // Routes
 router.get("/", async (_, res,next) => {
@@ -54,6 +31,7 @@ router.get("/:contactId", async (req, res, next) => {
 
 router.post("/", async (req, res, next) => {
   try {
+    console.log(userSchema.validate(req.body));
     const {error} = userSchema.validate(req.body);
     if (error) {
       throw HttpError(400, error.details[0].message);
