+++ schemas/index.js
@@ -1,0 +1,5 @@
+const userSchema = require("./user-schema");
+
+    module.exports = {
+    userSchema,
+}
